Move partition key range definition from top level
