+++ sdk/cosmos/src/lib.rs
@@ -108,7 +108,6 @@ pub mod responses;
 mod consistency_level;
 mod errors;
 mod headers;
-mod partition_key_range;
 mod partition_keys;
 mod query;
 mod resource_quota;
@@ -116,7 +115,6 @@ mod to_json_vector;
 mod traits;
 
 pub use consistency_level::ConsistencyLevel;
-pub use partition_key_range::PartitionKeyRange;
 pub use partition_keys::PartitionKeys;
 pub use query::{Param, ParamDef, Query};
 pub use resource_quota::ResourceQuota;
