+++ sdk/cosmos/src/responses/get_partition_key_ranges_response.rs
@@ -1,6 +1,5 @@
 use crate::headers::from_headers::*;
 use crate::CosmosError;
-use crate::PartitionKeyRange;
 use azure_core::headers::session_token_from_headers;
 use chrono::{DateTime, Utc};
 use http::response::Response;
@@ -72,4 +71,26 @@ impl std::convert::TryFrom<Response<Vec<u8>>> for GetPartitionKeyRangesResponse 
             partition_key_ranges: r.partition_key_ranges,
         })
     }
+}
+
+#[derive(Debug, PartialEq, PartialOrd, Clone, Deserialize)]
+pub struct PartitionKeyRange {
+    #[serde(rename = "_rid")]
+    pub rid: String,
+    pub id: String,
+    #[serde(rename = "_etag")]
+    pub etag: String,
+    #[serde(rename = "minInclusive")]
+    pub min_exclusive: String,
+    #[serde(rename = "maxExclusive")]
+    pub max_exclusive: String,
+    #[serde(rename = "ridPrefix")]
+    pub rid_prefix: u64,
+    pub _self: String,
+    #[serde(rename = "throughputFraction")]
+    pub throughput_fraction: u64,
+    pub status: String,
+    // TODO: parents
+    #[serde(rename = "_ts")]
+    pub ts: u64,
 }
